Update index.rst
The current Audience section undermines the power the CMF can provide to custom applications. If we just wanted a few simple content pages in a custom Symfony2 app, the CMF would be an overkill. It's important to point out that the CMF allows custom Symfony2 apps to inherit enterprise content management features.