Changed paths from bin/console to app/console
The bin/console does not exist after running the above getting started commands so it is changed to app/console. Also added --force to `php app/console doctrine:phpcr:init:dbal --force` because of the notice you get otherwise: `ATTENTION: This operation should not be executed in a production environment. Please use "--force" to execute the command.`