Document necessary App document mapping in intro
As part of a fix of a hard-coded mapping, default application level mapping is no longer offered out of the box. In a minimal setup, mappings section is now necessary in most cases to offer document mappings on application level. Since the hard-coded mapping was incorrect in the first place, this document change applied in all maintained versions. Update the documentation to reflect this change.